+++ www/go/modules/community/tasks/views/extjs3/TaskDetail.js
@@ -250,7 +250,8 @@ go.modules.community.tasks.TaskDetail = Ext.extend(go.detail.Panel, {
 
 	onLoad: function () {
 		this.getTopToolbar().getComponent("edit").setDisabled(this.data.permissionLevel < go.permissionLevels.write);
-		this.deleteItem.setDisabled(this.data.permissionLevel < go.permissionLevels.writeAndDelete);
+		// this.deleteItem.setDisabled(this.data.permissionLevel < go.permissionLevels.writeAndDelete);
+		this.deleteItem.setVisible(this.data.permissionLevel >= go.permissionLevels.writeAndDelete);
 
 		this.assignMeBtn.setVisible(!this.data.responsibleUserId);