+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Finance: Put invoice attachments before template attachments
 - Core: Fixed AutocompleteField not submitting when emptied because it didn't know it was modified if empty.
 - Files: hide portlets if user has no main access
+- Tasks / Support: hide delete button if permissions do not allow deleting a task
+- Tasks / Support: hide tasklist context menu items if user has no manage task lists permissions
 - Core: Fixed slow browser problem when making a lot of changes to the User (by selecting task lists or note books etc.)
 - Support / Supportclient: Focus editor when opening ticket